requirements.txt
Django==6.0.1
asgiref==3.11.0
sqlparse==0.5.5
tzdata==2025.3
pillow==12.1.0
django-ckeditor==6.7.3
django-js-asset==3.1.2
python-dotenv==1.2.1
django-admin-sortable2==2.3.1